+++ fda-recall-monitor/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.0.6
+
+- Fix bug when recall number was "N/A", resulting in collisions
+- Improve Google search query for API results
+
 ## 0.0.5
 
 - Fix an issue where new API results would not appear before acknowledged page results